test: 테스트 파일 통합
